backend/pyproject.toml
[project]
name = "debate-agent"
version = "0.1.0"
description = "Multi-agent debate system for decision making"
requires-python = ">=3.11"
dependencies = [
    "camel-ai>=0.2.0",
    "mem0ai>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
    "openai>=1.12.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"